requirements.txt
python-dotenv==1.1.0
pymongo==4.11.3
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.0
google-generativeai==0.8.4
json5==0.11.0